Added testing code
