+++ src/testing/requirements.txt
@@ -1,1 +1,2 @@
 requests
+pytest
