+++ src/testing/test_wrong.json
@@ -1,0 +1,1 @@
+{"example": 123}
